% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stops.R
\name{get_stops}
\alias{get_stops}
\title{Get a list of TFE stops, with locations and services calling}
\usage{
get_stops()
}
\value{
A tibble
}
\description{
Get a list of TFE stops, with locations and services calling
}
\examples{
get_stops()
}
