% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_routes.R
\name{list_routes}
\alias{list_routes}
\title{List all TFE routes}
\usage{
list_routes(.details = FALSE)
}
\arguments{
\item{.details}{Show route details?
These are nested as lists within the tibble.}
}
\value{
A tibble
}
\description{
Lists all Lothian Buses, Airlink, Skylink,
Lothian Country, East Coast Buses, Nightbus,
Nighthawk, and Edinburgh Trams services
}
\examples{
list_routes()
}
