% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_services.R
\name{list_services}
\alias{list_services}
\title{List all TFE services}
\usage{
list_services()
}
\value{
A tibble
}
\description{
Lists all Lothian Buses, Airlink, Skylink,
Lothian Country, East Coast Buses, Nightbus,
Nighthawk, and Edinburgh Trams services
}
\examples{
list_services()
}
