% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/live_departures.R
\name{live_departures}
\alias{live_departures}
\title{Get TfE live departures}
\usage{
live_departures(
  stop_code,
  api = "https://tfe-opendata.com/api/",
  version = 1,
  show_bus = FALSE
)
}
\arguments{
\item{stop_code}{SMS code for bus stop or vector of SMS codes}

\item{api}{TfE API endpoint}

\item{version}{TfE API version}
}
\value{
A tibble with next departures from selected stops
}
\description{
Get TfE live departures
}
\examples{
live_departures(c(36235979, 36253636))
}
