% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/live_locations.R
\name{live_locations}
\alias{live_locations}
\title{Live Lothian bus and Edinburgh tram locations}
\usage{
live_locations(api = "https://tfe-opendata.com/api/", version = 1)
}
\arguments{
\item{api}{TfE API endpoint}

\item{version}{TfE API version}
}
\value{
A tibble with live locations of all buses and trams
}
\description{
Live Lothian bus and Edinburgh tram locations
}
\examples{
live_locations()
}
